Fixing SDK-116 memory leak in ClipTag.
The ClipTag was hanging on to a CasInputStream.  This internally used a PipedInputStream with a 1MB buffer.  It turns out that this buffer is not released when you close the PipedInputStream.  At a higher level, failed transfers cause the CasSyncObject to be retained for reporting that will each hold one or more ClipTag objects and thus, the 1MB buffer, causing OutOfMemory errors after a couple thousand failed objects.